+++ MiningPlatform/apps/transaction-signer/src/main.ts
@@ -1,0 +1,11 @@
+/**
+ * MiningPlatform
+ * Author: Abia Nugrahanto
+ * Copyright (c) 2026 Abia Nugrahanto. All rights reserved.
+ */
+
+import { printVersionAndExitIfRequested } from '@mining/build-info';
+
+if (printVersionAndExitIfRequested('transaction-signer')) process.exit(0);
+
+void import('./runtime.js');